% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_flextable.R
\name{epi_flextable}
\alias{epi_flextable}
\title{Create publication-ready flextable}
\usage{
epi_flextable(model)
}
\arguments{
\item{model}{A fitted regression model}
}
\value{
A flextable object
}
\description{
Create publication-ready flextable
}
