% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tables.R
\name{epi_rep_table}
\alias{epi_rep_table}
\title{Create epidemiologic regression results table}
\usage{
epi_rep_table(model, digits = 2, include_intercept = FALSE)
}
\arguments{
\item{model}{A fitted regression model}

\item{digits}{Number of decimal places}

\item{include_intercept}{Logical; include intercept row}
}
\value{
A formatted tibble
}
\description{
Create epidemiologic regression results table
}
